impl-submit-report - move EncryptedBarkReportData into bark/models
