+++ src/lib/bark/models/encrypted-bark-report-data.ts
@@ -1,0 +1,8 @@
+import { BarkReportData } from "./bark-report-data";
+
+export type EncryptedBarkReportData = Omit<
+  BarkReportData,
+  "ineligibilityReason"
+> & {
+  encryptedIneligibilityReason: string;
+};
